+++ innopoints/views/__init__.py
@@ -1,3 +1,4 @@
 from .authentication import *
+from .project import *
 
 from .views import *  # temporary, until all views have been moved
